ISS04 - initialize the first test suite
