+++ playground/test_case_01/test_case_01.py
@@ -1,0 +1,10 @@
+import unittest
+
+
+class TestCase01(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, True)
+
+
+if __name__ == '__main__':
+    unittest.main()
