--- python_src/lgsvl/dreamview/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2019-2020 LG Electronics, Inc.
-#
-# This software contains code licensed as described in LICENSE.
-#
-
-from .dreamview import Connection, CoordType
