--- python_src/lgsvl/evaluator/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2020 LG Electronics, Inc.
-#
-# This software contains code licensed as described in LICENSE.
-#
-
-from .utils import TestException, right_lane_check, separation, almost_equal, in_parking_zone
