--- python_src/source/evaluator/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2019 LG Electronics, Inc.
-#
-# This software contains code licensed as described in LICENSE.
-#
-
-from .utils import TestException, separation
