--- python_src/source/simulator/__init__.py
@@ -1,1 +1,0 @@
-from .simulator import Simulator
